[CHORE] Delete Hayseed Hank Chores (Yes from 2 years ago)
