+++ src/features/island/delivery/components/Orders.tsx
@@ -26,6 +26,7 @@ import { NPCName, NPC_WEARABLES } from "lib/npcs";
 import {
   getDayOfYear,
   millisecondsToString,
+  secondsTillReset,
   secondsToString,
 } from "lib/utils/time";
 import {
@@ -41,7 +42,6 @@ import { RequirementLabel } from "components/ui/RequirementsLabel";
 import { Button } from "components/ui/Button";
 import { ButtonPanel, InnerPanel, OuterPanel } from "components/ui/Panel";
 import { getSeasonalTicket } from "features/game/types/seasons";
-import { secondsTillReset } from "features/helios/components/hayseedHank/HayseedHankV2";
 import { Revealing } from "features/game/components/Revealing";
 import { Revealed } from "features/game/components/Revealed";
 import { Label } from "components/ui/Label";
