+++ src/features/world/ui/player/PlayerGift.tsx
@@ -4,9 +4,8 @@ import { SUNNYSIDE } from "assets/sunnyside";
 import { Button } from "components/ui/Button";
 import { Label } from "components/ui/Label";
 import { Revealed } from "features/game/components/Revealed";
-import { secondsTillReset } from "features/helios/components/hayseedHank/HayseedHankV2";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import { secondsToString } from "lib/utils/time";
+import { secondsTillReset, secondsToString } from "lib/utils/time";
 import { useContext, useState } from "react";
 import { ChestRevealing } from "../chests/ChestRevealing";
 import { Context } from "features/game/GameProvider";
